version: change all fields to uint16_t
Ok, shame on me: the version string was so obviously correct that I only
verified that the comparisons were working as expected.

Turns out it isn't: http://lists.boost.org/boost-users/2006/12/24194.php

boost::format will treat uint8_t arguments as char, and therefore we will end
up with the version string misprinted.

We can just cast it to uint16_t before we print, but since this is not exactly
a struct that we will be using all the time, let's favor readability over
saving a few bytes, and change all fields to uint16_t.

+++ version.hh
@@ -6,9 +6,9 @@
 
 namespace version {
 class version {
-    std::tuple<uint8_t, uint8_t, uint16_t> _version;
+    std::tuple<uint16_t, uint16_t, uint16_t> _version;
 public:
-    version(uint8_t x, uint8_t y = 0, uint16_t z = 0): _version(std::make_tuple(x, y, z)) {}
+    version(uint16_t x, uint16_t y = 0, uint16_t z = 0): _version(std::make_tuple(x, y, z)) {}
 
     sstring to_sstring() {
         return sprint("%d.%d.%d", std::get<0>(_version), std::get<1>(_version), std::get<2>(_version));
